Get Historical Volatility and Attributes of Stocks
This workbook will ask you for stock tickers when it opens. It will then pull the attributes of the tickers and ask for a date range to measure the historical volatility of the stock tickers. The end result displays a correlation matrix, historical volatility, and attributes of the stock. You can copy this code and put it into a module. 

Add the following worksheet names:

	Test
	Volatility
	Data
	Matrix

In the workbook module add the following:

Private Sub Workbook_Open()
RunMe
End Sub

Close workbook and open to run the macro.